refactor(core): split validation rules into focused modules
